requirements.txt
pybit>=5.6.0
python-dotenv>=1.0.0
requests>=2.31.0
hyperliquid-python-sdk>=0.4.0
